"Update permission check in Influactive Modal plugin
Updated the user permission check function in Influactive Modal plugin. Previously, the plugin checked if the user has the 'manage_options' permission. Now it checks if the user has the 'edit_posts' permission. This change is to ensure that only users with the appropriate permissions are able to edit posts, enhancing security."