Update plugin version to 1.1
Updated the 'Modal by Influactive' plugin to version 1.1. The version number is updated in the plugin PHP file, CSS, JavaScript, and language files. This is to ensure the version number aligns with the recently made changes in the plugin, thus allowing users to know they are using the latest version.