Add null coalescing operator to plugin options
Updated modal title and content retrieval in 'Modal by Influactive' plugin to prevent potential null values. The null coalescing operator '??' is now used during the 'get_option' calls. This way, in the case of the absence of a title or content, an empty string will be returned, preventing possible null-related errors when displaying the modal.